refactor: split public and protected smoke checks
